requirements/base.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-environ==0.11.2
psycopg2-binary==2.9.10
Pillow==11.1.0
celery==5.4.0
redis==5.2.1
django-redis==5.4.0
celery[redis]==5.4.0
django-filter==24.3
python-slugify==8.0.4
pywebpush==2.0.0
